链表实现，消息18和消息24A ，未加消息丢弃 LISTVIEW界面 AllList 正常，MoniteSetting船名还有问题监控还好。

+++ BoardSupport/Drawinformation.c
@@ -19,7 +19,7 @@ float getDistance(const long lt, const long lg)
 	
 }
 
-int strtoi(const char*str)
+long strtoi(const char*str)
 {
 	int val = 0;
 	short i = 0;
